moves collocate work to compare, removes references to collocate
