--- app/assets/scripts/views/collocate/index.js
@@ -1,3 +1,0 @@
-import Collocate from './collocate';
-
-export default Collocate;
