--- app/assets/styles/collocate/_content.scss
@@ -1,233 +1,0 @@
-/* ==========================================================================
-   Collocate: Content
-   ========================================================================== */
-
-   .page--collocate {
-    .inpage__header {
-      background: tint($base-color, 96%);
-    }
-  
-    .inpage__headline {
-      @include media(large-up) {
-        grid-column: 1 / span 8;
-      }
-    }
-  
-    .fold > .inner {
-      grid-template-columns: 1fr;
-    }
-
-    .choose-sensors-wrapper {
-      padding: 6rem 0 3rem;
-    }
-
-    .refresh-sensors {
-      margin: 0.25rem 0;
-    }
-
-    .collocation-sensors-list {
-      box-shadow: inset 0px -3px 4px rgba(0, 0, 0, 0.25);
-      margin-top: 0.6rem;
-      background-color: #efefef;
-      padding: 0.25rem 1.2rem;
-      max-width: 400px;
-      max-height: 80px;
-      overflow-y: scroll;
-      margin-right: 3%;
-      position: absolute;
-      float: left;
-      transition: max-height 0.15s ease-in;
-      p {
-        margin: 0;
-      }
-      &:hover {
-        box-shadow: 0px 3px 4px rgba(0, 0, 0, 0.25);
-        height: inherit;
-        max-height: 600px;
-        transition: max-height 0.15s ease-out;
-      }
-    }
-
-    .location_selector {
-      h4 {
-        margin-bottom: 0.25rem;
-      }
-      padding-bottom: 1rem;
-    }
-
-    .location_selector_spatial {
-      padding: 0;
-      margin-top: 0;
-    }
-
-    .collocate-sensor {
-      margin: 0.2rem 0;
-      cursor: pointer;
-      span {
-        font-weight: bold;
-      }
-    }
-
-    .collocate-map {
-      padding-top: 0;
-    }
-  }
-
-  .search-back-button .button {
-    padding: 0.5rem 0;
-  }
-  
-  %list-reset {
-    margin: 0;
-    padding: 0;
-    list-style: none;
-  }
-  
-  .collocate {
-    grid-row: 2;
-    grid-column: 1 / span 12;
-  
-    &__location-list {
-      @extend %list-reset;
-      overflow: hidden;
-    }
-  
-    &__location {
-      @include col(12/12);
-      padding-left: 2rem;
-      margin-top: 2rem;
-  
-      &:first-child {
-        margin-top: 0;
-      }
-  
-      @include media(small-up) {
-        @include col(6/12, $cycle: 2);
-        &:nth-child(2) {
-          margin-top: 0;
-        }
-      }
-      @include media(medium-up) {
-        @include col(4/12, $cycle: 3);
-        &:nth-child(3) {
-          margin-top: 0;
-        }
-      }
-  
-      &--actions {
-        @include col(12/12);
-        margin-top: 2rem;
-        padding-left: 0;
-  
-        @include media(medium-up) {
-          @include col(4/12);
-        }
-  
-        .form__actions .button {
-          margin-right: 0.5rem;
-  
-          &:last-child {
-            margin-right: 0;
-          }
-        }
-      }
-    }
-  
-    &__subtitle {
-      @extend .heading-alt;
-      font-size: 0.75rem;
-      line-height: 1.25rem;
-      margin-bottom: 0;
-  
-      @include media(medium-up) {
-        font-size: 0.875rem;
-      }
-    }
-  
-    &__title {
-      @extend .heading--small;
-      position: relative;
-      line-height: 1.75rem;
-      margin-bottom: 0.25rem;
-  
-      small {
-        display: block;
-      }
-  
-    }
-  
-    &__actions {
-      > [class*="button"]:first-child {
-        margin-left: - $global-spacing / 2;
-      }
-  
-      .button {
-        text-transform: uppercase;
-      }
-    }
-  
-    &__info-msg {
-      margin-top: -0.5rem;
-      p {
-        font-family: $base-font-family;
-        font-size: 0.875rem;
-        color: rgba($base-font-color, 0.48);
-        margin-bottom: 0.5rem;
-      }
-    }
-  }
-  
-  .collocate-parameters {
-    font-size: 0.875rem;
-  }
-  
-  .collocate-marker {
-    display: inline-block;
-    &::before {
-      content: '';
-      position: absolute;
-      width: 1rem;
-      height: 1rem;
-      top: 0.325rem;
-      left: -2rem;
-      border-radius: $base-border-radius;
-      background: $base-color;
-  
-      @include media(medium-up) {
-        width: 1.25rem;
-        height: 1.25rem;
-        top: 0.25rem;
-      }
-    }
-  
-    &--st::before {
-      background-color: $infographic-color-a;
-    }
-  
-    &--nd::before {
-      background-color: $infographic-color-b;
-    }
-  
-    &--rd::before {
-      background-color: $infographic-color-c;
-    }
-  }
-  
-  .button-collocate-location {
-    @extend .button, .button--medium, .button--base-unbounded;
-    color: rgba($base-font-color, 0.80);
-    margin-left: -($global-spacing / 2);
-  
-    &::before {
-      @extend %openaq-ds-icon-plus;
-    }
-  }
-  
-  
-  /* Fold: Measurements
-     ========================================================================== */
-  
-  #collocate-fold-measurements {
-  
-  }
-
