+++ app/assets/styles/main.scss
@@ -86,13 +86,6 @@
 
    @import "projects/hub";
    
-
-/* Collocate
-   ========================================================================== */
-
-@import "collocate/content";
-
-
 /* Compare
    ========================================================================== */
